created controllers for each pageType
